chore(store): move api call to wallet store
