+++ shared/store/wallet.ts
@@ -1,7 +1,7 @@
 
 import { WhaleApiClient } from '@defichain/whale-api-client'
 import { AddressToken } from '@defichain/whale-api-client/dist/api/address'
-import { PoolPairData } from '@defichain/whale-api-client/dist/api/poolpairs'
+import { AllSwappableTokensResult, PoolPairData } from '@defichain/whale-api-client/dist/api/poolpairs'
 import { TokenData } from '@defichain/whale-api-client/dist/api/tokens'
 import { createAsyncThunk, createSelector, createSlice, PayloadAction } from '@reduxjs/toolkit'
 import BigNumber from 'bignumber.js'
@@ -9,13 +9,19 @@ import BigNumber from 'bignumber.js'
 interface AssociatedToken {
   [key: string]: TokenData
 }
+export interface SwappableTokens {
+  [key: string]: AllSwappableTokensResult
+}
+
 export interface WalletState {
   utxoBalance: string
   tokens: WalletToken[]
   allTokens: AssociatedToken
   poolpairs: DexItem[]
+  swappableTokens: SwappableTokens
   hasFetchedPoolpairData: boolean
   hasFetchedToken: boolean
+  hasFetchedSwappableTokens: boolean
 }
 
 export interface WalletToken extends AddressToken {
@@ -32,6 +38,8 @@ const initialState: WalletState = {
   tokens: [],
   allTokens: {},
   poolpairs: [],
+  swappableTokens: {},
+  hasFetchedSwappableTokens: false,
   hasFetchedPoolpairData: false,
   hasFetchedToken: false
 }
@@ -106,6 +114,16 @@ export const fetchTokens = createAsyncThunk(
   }
 )
 
+export const fetchSwappableTokens = createAsyncThunk(
+  'wallet/swappableTokens',
+  async ({ client, fromTokenId }: {
+    client: WhaleApiClient
+    fromTokenId: string
+  }): Promise<AllSwappableTokensResult> => {
+    return await client.poolpairs.getSwappableTokens(fromTokenId)
+  }
+)
+
 export const wallet = createSlice({
   name: 'wallet',
   initialState,
@@ -125,6 +143,15 @@ export const wallet = createSlice({
       state.utxoBalance = action.payload.utxoBalance
       state.allTokens = associateTokens(action.payload.allTokens)
     })
+    builder.addCase(fetchSwappableTokens.fulfilled, (state, action: PayloadAction<AllSwappableTokensResult>) => {
+      state.hasFetchedSwappableTokens = true
+      state.swappableTokens = {
+        ...state.swappableTokens,
+        ...{
+          [action.payload.fromToken.id]: action.payload
+        }
+      }
+    })
   }
 })
 
@@ -143,17 +170,17 @@ const rawTokensSelector = createSelector((state: WalletState) => state.tokens, (
 })
 
 export const tokensSelector = createSelector([rawTokensSelector, (state: WalletState) => state.utxoBalance], (tokens, utxoBalance) => {
-    const utxoAmount = new BigNumber(utxoBalance)
-    const tokenAmount = new BigNumber((tokens.find(t => t.id === '0') ?? tokenDFI).amount)
-    return tokens.map((t) => {
-      if (t.id === '0_utxo') {
-        return { ...t, amount: utxoAmount.toFixed(8) }
-      } else if (t.id === '0_unified') {
-        return { ...t, amount: utxoAmount.plus(tokenAmount).toFixed(8) }
-      }
-      return t
-    })
-  }
+  const utxoAmount = new BigNumber(utxoBalance)
+  const tokenAmount = new BigNumber((tokens.find(t => t.id === '0') ?? tokenDFI).amount)
+  return tokens.map((t) => {
+    if (t.id === '0_utxo') {
+      return { ...t, amount: utxoAmount.toFixed(8) }
+    } else if (t.id === '0_unified') {
+      return { ...t, amount: utxoAmount.plus(tokenAmount).toFixed(8) }
+    }
+    return t
+  })
+}
 )
 
 export const DFITokenSelector = createSelector(tokensSelector, tokens => {
@@ -186,6 +213,6 @@ export const tokenSelector = createSelector([tokensSelector, selectTokenId], (to
 /**
  * Get single token detail by `displaySymbol` from wallet store.
  */
- export const tokenSelectorByDisplaySymbol = createSelector([(state: WalletState) => state.allTokens, selectTokenId], (allTokens, displaySymbol) => {
+export const tokenSelectorByDisplaySymbol = createSelector([(state: WalletState) => state.allTokens, selectTokenId], (allTokens, displaySymbol) => {
   return allTokens[displaySymbol]
 })
